[Security Solution] Relocate sourcerer_paths to data_view_manager/utils
The route->PageScope and data view picker visibility helpers are pure
data view manager logic (they import PageScope from data_view_manager),
so sourcerer/containers/sourcerer_paths.ts moves to
data_view_manager/utils/paths.ts. Consumer imports updated accordingly.
