+++ x-pack/solutions/security/plugins/security_solution/public/app/components/top_values_popover/top_values_popover.tsx
@@ -11,7 +11,7 @@ import { i18n } from '@kbn/i18n';
 import { useLocation } from 'react-router-dom';
 import useObservable from 'react-use/lib/useObservable';
 import { StatefulTopN } from '../../../common/components/top_n';
-import { getScopeFromPath } from '../../../sourcerer/containers/sourcerer_paths';
+import { getScopeFromPath } from '../../../data_view_manager/utils/paths';
 import { useKibana } from '../../../common/lib/kibana';
 import { useDataView } from '../../../data_view_manager/hooks/use_data_view';
 import { useBrowserFields } from '../../../data_view_manager/hooks/use_browser_fields';
