+++ x-pack/solutions/security/plugins/security_solution/public/app/home/global_header/index.test.tsx
@@ -16,7 +16,7 @@ import {
 } from '../../../../common/constants';
 import { createMockStore, mockGlobalState, TestProviders } from '../../../common/mock';
 import { TimelineId } from '../../../../common/types/timeline';
-import { dataViewPickerPaths } from '../../../sourcerer/containers/sourcerer_paths';
+import { dataViewPickerPaths } from '../../../data_view_manager/utils/paths';
 
 import { DATA_VIEW_PICKER_TEST_ID } from '../../../data_view_manager/components/data_view_picker/constants';
 import { useKibana as mockUseKibana } from '../../../common/lib/kibana/__mocks__';
